Significant refactor; removed OpenAPI integration and curly braces definitino for params, now only using ExpressJS ':' params
