+++ src/index.ts
@@ -1,45 +1,4 @@
-import zemi from "./core";
+import Zemi from "./core";
 
-export default zemi;
-export { default as ZemiOpenApiSpecGenerator } from "./openapi";
-
-export {
-  ZemiMethod,
-  ZemiOpenApiSpecGenerationOptions,
-  ZemiHandlerDefinition,
-  ZemiRequest,
-  ZemiRequestHandler,
-  ZemiResponse,
-  ZemiRoute,
-  ZemiRouteDefinition,
-} from "./types/core.types";
-
-export {
-  OpenApiReferenceObject,
-  OpenApiCallbackObject,
-  OpenApiSecurityRequirementObject,
-  OpenApiParameterObjectIn,
-  OpenApiPathItemObject,
-  OpenApiDiscriminatorObject,
-  OpenApiXmlObject,
-  OpenApiSchemaObject,
-  OpenApiHeaderObject,
-  OpenApiEncodingObject,
-  OpenApiMediaTypeObject,
-  OpenApiRequestBodyObject,
-  OpenApiParameterObject,
-  OpenApiLinkObject,
-  OpenApiResponseObject,
-  OpenApiServerVariableObject,
-  OpenApiServerObject,
-  OpenApiOAuthFlowObject,
-  OpenApiOAuthFlowsObject,
-  OpenApiSecuritySchemeObject,
-  OpenApiOperationObject,
-  OpenApiPathItemDefinitionObject,
-  OpenApiExampleObject,
-  OpenApiExternalDocumentationObject,
-  OpenApiInfoObject,
-  OpenApiTagsObject,
-  OpenApiDoc,
-} from "./types/openapi.types";
+export * from "./types";
+export default Zemi;
