--- src/types/helpers.types.ts
@@ -1,4 +1,0 @@
-export interface NamedRoute {
-  name: string;
-  path: string;
-}
